Change `repo-branch: master` to `main` for urls
### Problem

The branch name of the repo is main but urls are created according to the default branch name master as mentioned at quarto-dev/quarto-cli#4826. Because Edit this page and View source buttons in every page contains branch names in their urls, they do not work now.

### Solution

It updates `repo-branch: master` to `repo-branch: main` to solve this.